requirement.txt
torch
transformers
sentence-transformers
faiss-cpu
langchain
huggingface-hub
bitsandbytes
accelerate